Fix alignment of time text on passing --time_format
Time text shifted to the left and goes out of the box even
Cutted out from the screen if passes different `strftime`
like "%I:%M:S %p" to the -tf option

default alignment was `center` now is `left`. So from now text
expands to the right, May be cutted out of the box but not from
the screen at least.